+++ KHOBilgiSistemiMVCCoreWebApp/Areas/AkademikDanismanArea/Models/OgrenciDegerlendirmeVM/OgrenciDegerlendirmeKayit.cs
@@ -3,11 +3,12 @@
     public class OgrenciDegerlendirmeKayit
     {
         public int DegTurID { get; set; }
-        public int OgrenciID { get; set; }
-        public int PerID { get; set; }
+        public long OgrenciID { get; set; }
+        public long PerID { get; set; }
         public DateTime TarihSaat { get; set; }
         public string? Degerlendirme { get; set; }
         public int EOYiliID { get; set; }
         public int Donem { get; set; }
+      
     }
 }
